DTE: Don't flush the inqueues on disconnect
Previously, the inqueues on both sides were flushed on disconnect by the
dying EventSender and EventReceiver. However, this introduces
indeterminacy as it is concurrent with the EventReplayer taking events
from the inqueue. Remove the flushing and let the EventReplayer empty
the queue in its usual way. It must be ensured that there won't arrive
any events from a particular editor after a DisconnectEvent, but that is
not difficult since only the EventReceivers can put events in the
inqueues and the EventSenders don't send out any more events after
having sent a DisconnectEvent.

We might consider changing this queueing stuff into an observer thing.