Stop new-client listener on server
Previously, we let this thread running because we thought that after the
server disconnected, the process would stop and the JVM would take care
of the socket. However, it might happen that we want to disconnect the
editor with the server and then connect it again. When then the server
is moved to it, it would try to open a ServerSocket on the old port and
thus crash. Therefore add an ugly timeout thing that periodically checks
if the listener thread has been interrupted and, if yes, stops it. –
There appears to be no sane way to do this in Java.