Fix event buffering during server moving
The intent behind buffering TextChangeEvents from user input during
server moving (!) was, that the user could type and type and type and
type and wouldn't notice if the server was moved. However, I buffered
the TextChangeEvents from the inqueue of the ClientEventDistributor so
that they didn't reach the EventDisplayer. So what the user typed
appeared on the screen, but only after the server moving was over.
Change this, so that the events are given to the EventDisplayer, but
only sent after the new server is in place.