Add callbacks for getting listen information and thereby fix server move
Previously the server was always started with localhost as its listen
address. This of course broke the server moving with non-local editors
participating, since the clients got 0.0.0.0/0.0.0.0 to connect to.

Therefore add callbacks to actually use the information from the
_listenIp text field. And change some other stuff that became obsolete
or was wrong. – Should probably have made two out of this.